Refactor size_t (#11)
* refactoring unsigned int, et.al. using size_t

* fixed type of millis variable

* refactoring using clang-format, fixed casting, fixed alignment of _address[4]

* strict / pedantic checking resulting in error (-Werror) 

* fixed potential buffer overflow, fixed cast warnings

* added missing function declarations in tests

* added Arduino Uno

* exclude esp8266 examples from Arduino Uno

* updated callback signature in tests and examples

* added -Wnull-dereference, -Wreorder in tests

* using static_cast<unsigned long>, optimized buildHeader(), size_t is wrong in writeString()

+++ src/PubSubClient.h
@@ -111,9 +111,9 @@
  * @brief Define the signature required by any callback function.
  * @note The parameters are TOPIC, PAYLOAD, and LENGTH, respectively.
  */
-#define MQTT_CALLBACK_SIGNATURE std::function<void(char*, uint8_t*, unsigned int)> callback
+#define MQTT_CALLBACK_SIGNATURE std::function<void(char*, uint8_t*, size_t)> callback
 #else
-#define MQTT_CALLBACK_SIGNATURE void (*callback)(char*, uint8_t*, unsigned int)
+#define MQTT_CALLBACK_SIGNATURE void (*callback)(char*, uint8_t*, size_t)
 #endif
 
 #define CHECK_STRING_LENGTH(l, s)                                  \
@@ -136,7 +136,7 @@ class PubSubClient : public Print {
    private:
     Client* _client;
     uint8_t* buffer;
-    uint16_t bufferSize;
+    size_t bufferSize;
     uint16_t keepAlive;
     uint16_t socketTimeout;
     uint16_t nextMsgId;
@@ -147,13 +147,13 @@ class PubSubClient : public Print {
     uint32_t readPacket(uint8_t*);
     bool readByte(uint8_t* result);
     bool readByte(uint8_t* result, uint16_t* index);
-    bool write(uint8_t header, uint8_t* buf, uint16_t length);
-    uint16_t writeString(const char* string, uint8_t* buf, uint16_t pos);
+    bool write(uint8_t header, uint8_t* buf, size_t length);
+    size_t writeString(const char* string, uint8_t* buf, size_t pos);
     // Build up the header ready to send
     // Returns the size of the header
     // Note: the header is built at the end of the first MQTT_MAX_HEADER_SIZE bytes, so will start
     //       (MQTT_MAX_HEADER_SIZE - <returned size>) bytes into the buffer
-    size_t buildHeader(uint8_t header, uint8_t* buf, uint16_t length);
+    size_t buildHeader(uint8_t header, uint8_t* buf, size_t length);
     IPAddress ip;
     char* domain;
     uint16_t port;
@@ -367,13 +367,13 @@ class PubSubClient : public Print {
      * @return true If the buffer was resized.
      * false If the buffer could not be resized.
      */
-    bool setBufferSize(uint16_t size);
+    bool setBufferSize(size_t size);
 
     /**
      * @brief Gets the current size of the internal buffer.
      * @return The size of the internal buffer.
      */
-    uint16_t getBufferSize();
+    size_t getBufferSize();
 
     /**
      * @brief Connects the client.
@@ -464,7 +464,7 @@ class PubSubClient : public Print {
      * @return true If the publish succeeded.
      * false If the publish failed, either connection lost or message too large.
      */
-    bool publish(const char* topic, const uint8_t* payload, unsigned int plength);
+    bool publish(const char* topic, const uint8_t* payload, size_t plength);
 
     /**
      * @brief Publishes a message to the specified topic.
@@ -475,7 +475,7 @@ class PubSubClient : public Print {
      * @return true If the publish succeeded.
      * false If the publish failed, either connection lost or message too large.
      */
-    bool publish(const char* topic, const uint8_t* payload, unsigned int plength, bool retained);
+    bool publish(const char* topic, const uint8_t* payload, size_t plength, bool retained);
 
     /**
      * @brief Publishes a message stored in PROGMEM to the specified topic.
@@ -496,7 +496,7 @@ class PubSubClient : public Print {
      * @return true If the publish succeeded.
      * false If the publish failed, either connection lost or message too large.
      */
-    bool publish_P(const char* topic, const uint8_t* payload, unsigned int plength, bool retained);
+    bool publish_P(const char* topic, const uint8_t* payload, size_t plength, bool retained);
 
     /**
      * @brief Start to publish a message.
@@ -512,7 +512,7 @@ class PubSubClient : public Print {
      * @return true If the publish succeeded.
      * false If the publish failed, either connection lost or message too large.
      */
-    bool beginPublish(const char* topic, unsigned int plength, bool retained);
+    bool beginPublish(const char* topic, size_t plength, bool retained);
 
     /**
      * @brief Finish sending a message that was started with a call to beginPublish.
